Animated home page title and titles on mobile versions of home and careers pages
