+++ src/components/home/desktop/HomeTitleBox.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import {Grid, makeStyles, styled, Typography,} from '@material-ui/core'
 import { theme } from "../../../theme";
+import {FadeInOutAnimation} from "../../FadeInOutAnimation";
 
 
 const StyledGrid = styled(Grid)({
@@ -23,18 +24,6 @@ const TitleTextPrimary = styled(Typography)({
   color: theme.palette.text.primary
 });
 
-const TitleTextSecondary = styled(Typography)({
-  fontFamily: theme.typography.fontFamily,
-  fontSize: '1.75vw',
-  fontWeight: 'bold',
-  fontStretch: "normal",
-  fontStyle: "normal",
-  lineHeight: 1,
-  letterSpacing: '-0.84px',
-  textAlign: "left",
-  color: theme.palette.text.secondary
-});
-
 const SubtitleText = styled(Typography)({
   marginTop: '0.813vw',
   fontFamily: theme.typography.fontFamily,
@@ -50,7 +39,7 @@ const SubtitleText = styled(Typography)({
 
 interface Props {
   titleTextPrimary: string;
-  titleTextSecondary: string;
+  titleTextSecondary: string[];
   subTitleText: string;
   titleColorPrimary?: string;
   titleColorSecondary?: string;
@@ -59,12 +48,20 @@ interface Props {
 
 export const HomeTitleBox: React.FC<Props> = (props: Props) => {
 
-  const colors = makeStyles({
+  const styles = makeStyles({
     textPrimary: {
       color: props.titleColorPrimary ? props.titleColorPrimary : theme.palette.text.primary
     },
     textSecondary: {
-      color: props.titleColorSecondary ? props.titleColorSecondary : theme.palette.text.secondary
+      color: props.titleColorSecondary ? props.titleColorSecondary : theme.palette.text.secondary,
+      fontFamily: theme.typography.fontFamily,
+      fontSize: '1.75vw',
+      fontWeight: 'bold',
+      fontStretch: "normal",
+      fontStyle: "normal",
+      lineHeight: 1,
+      letterSpacing: '-0.84px',
+      textAlign: "left",
     }
   })();
 
@@ -72,14 +69,14 @@ export const HomeTitleBox: React.FC<Props> = (props: Props) => {
     <StyledGrid className={props.classes} container direction={'column'} spacing={0} justify={'flex-start'} alignItems={'flex-start'}>
       <Grid item container direction={'row'} spacing={0} justify={'flex-start'} alignItems={'flex-start'}>
         <Grid item>
-          <TitleTextPrimary className={colors.textPrimary}>{props.titleTextPrimary}</TitleTextPrimary>
+          <TitleTextPrimary className={styles.textPrimary}>{props.titleTextPrimary}</TitleTextPrimary>
         </Grid>
         <Grid item>
-          <TitleTextSecondary className={colors.textSecondary}>&nbsp;{props.titleTextSecondary}</TitleTextSecondary>
+          <FadeInOutAnimation textSecondary={props.titleTextSecondary} textClasses={styles.textSecondary} />
         </Grid>
       </Grid>
       <Grid item>
-        <SubtitleText className={colors.textPrimary}>{props.subTitleText}</SubtitleText>
+        <SubtitleText className={styles.textPrimary}>{props.subTitleText}</SubtitleText>
       </Grid>
     </StyledGrid>
   );
