+++ src/components/home/mobile/HomeTItleBoxMobile.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
-import {Grid, styled, Typography,} from '@material-ui/core'
+import {Grid, makeStyles, styled, Typography,} from '@material-ui/core'
 import { theme } from "../../../theme";
+import {FadeInOutAnimation} from "../../FadeInOutAnimation";
 
 
 const StyledGrid = styled(Grid)({
@@ -23,19 +24,6 @@ const TitleTextPrimary = styled(Typography)({
   color: theme.palette.text.primary
 });
 
-const TitleTextSecondary = styled(Typography)({
-  marginTop: '0.967vw',
-  fontFamily: theme.typography.fontFamily,
-  fontSize: '6.67vw',
-  fontWeight: 'bold',
-  fontStretch: "normal",
-  fontStyle: "normal",
-  lineHeight: 1,
-  letterSpacing: '-0.84px',
-  textAlign: "left",
-  color: theme.palette.text.secondary
-});
-
 const SubtitleText = styled(Typography)({
   marginTop: '3.6vw',
   fontFamily: theme.typography.fontFamily,
@@ -51,19 +39,35 @@ const SubtitleText = styled(Typography)({
 
 interface Props {
   titleTextPrimary: string;
-  titleTextSecondary: string;
+  titleTextSecondary: string[];
   subTitleText: string;
   classes?: string;
 }
 
 export const HomeTitleBoxMobile: React.FC<Props> = (props: Props) => {
+
+  const styles = makeStyles({
+    textSecondary: {
+      marginTop: '0.967vw',
+      fontFamily: theme.typography.fontFamily,
+      fontSize: '6.67vw',
+      fontWeight: 'bold',
+      fontStretch: "normal",
+      fontStyle: "normal",
+      lineHeight: 1,
+      letterSpacing: '-0.84px',
+      textAlign: "left",
+      color: theme.palette.text.secondary
+    }
+  })();
+
   return (
     <StyledGrid className={props.classes} container direction={'column'} spacing={0} justify={'flex-start'} alignItems={'flex-start'}>
       <Grid item>
         <TitleTextPrimary>{props.titleTextPrimary}</TitleTextPrimary>
       </Grid>
       <Grid item>
-        <TitleTextSecondary>{props.titleTextSecondary}</TitleTextSecondary>
+        <FadeInOutAnimation noSpace textSecondary={props.titleTextSecondary} textClasses={styles.textSecondary} />
       </Grid>
       <Grid item>
         <SubtitleText>{props.subTitleText}</SubtitleText>
